requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.19.0
scikit-learn==1.4.1.post1
psycopg2-binary==2.9.9
bcrypt==4.1.2
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0
opencv-python==4.9.0.80
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
